Added vibe coded SPARQL Parsing Slop (Semi Functional, wont be embedded into the App. Added first domain specific backend Functions. Adjusted state response in both Code and Specs.)

+++ app/backend/semantics/src/main/java/com/github/jcodevandamme/semantics/app/dto/response/FactorResponse.java
@@ -1,4 +1,4 @@
 package com.github.jcodevandamme.semantics.app.dto.response;
 
-public record FactorResponse(float factor) {
+public record FactorResponse(double factor) {
 }
